+++ tests/ldap/conftest.py
@@ -4,6 +4,7 @@ from unittest.mock import MagicMock, Mock
 
 import pytest
 from ldap3 import Connection
+from ldap3.core.exceptions import LDAPSocketOpenError
 from pytest import MonkeyPatch
 
 from mex.common.ldap.connector import LDAPConnector
@@ -58,3 +59,14 @@ def ldap_mocker(monkeypatch: MonkeyPatch) -> LDAPMocker:
         monkeypatch.setattr(LDAPConnector, "__init__", __init__)
 
     return mocker
+
+
+@pytest.fixture(autouse=True)
+def skip_ldap_integration_tests_on_connection_error(
+    is_integration_test: bool,  # noqa: FBT001
+) -> None:
+    if is_integration_test:
+        try:
+            LDAPConnector.get()
+        except LDAPSocketOpenError:
+            pytest.skip("LDAP unavailable")