+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- don't require env settings for ldap and wiki tests
+
 ### Deprecated
 
 ### Removed
@@ -22,10 +24,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.5.0] - 2025-09-30
 
 ### Added
+
 - bumped pytz to newest version >=2025
 - BREAKING: add RKI organization as affiliation to ldap persons
 
 ### Fixed
+
 - fixed temporal entity tests for newest pytz version
 
 ## [1.4.0] - 2025-09-11